Container runs/builds should user a inner user
Saving every output from a build/test as root gives trouble for the
user. This tries to detect if the host is Linux (for now), create a
basic structure to run the build, and eventually `exec` and `run`, as a
`runner` user, and keeps every build pack installated files, shared
cache/output directories owned by it.

The runner user is mapped to the host user, using the same uid:gid.

Also cleans up some code around the defaults for a container, tries to
split some of the major tweak functions into another source file.

Aso introduces a rough method of identifying which commands are
considered to be setup commands, that should run as root (like
`apt-get`).

Splits up SSH forwarder injection bits into a second source file, and
also make it so that it doesn't always repeat itself.

Fixes: "As a CLI UNIX POSIX user I expect that build" [ch1780]