Remove progress spinners
The code to render the spinners was improperly copied into yb without
the correct licensing information: it should be added as a dependency or
properly sectioned off if we were to add it back.

It was only being used in remote build, and I'd like to revisit the
output in a more holistic take.