Add biome record/replay infrastructure
Buildpacks are tested by installing them into an isolated directory and
checking whether they run. With this change, rather than having to go
through a full install (which takes a long time), the commands executed
in the biome are executed against a previous known-good run for each
platform. The data files are generated by running `go test` with a
`-record` flag.

This required a few tweaks to the Android SDK buildpack (to make it
behave deterministically) and to the Heroku buildpack test (to opt out
of using replay), but otherwise the buildpack tests are identical.

The tests will still download full SDKs from upstream, which is slow,
but this keeps confidence that the URLs still work and haven't changed.

[Fixes ch2764]