Mount an eventually consistent HOME for Docker-on-Mac
On Docker-on-Mac, there's a translation layer for volumes to go from the
Linux VM running Docker to the macOS host folder. This introduces heavy
overhead for our case where the home directory is largely cache and
doesn't need to be strongly consistent on the host side.

We hint to Docker that we only want eventual consistency for this
volume, which means that writes to the volume from the container might
not be propagated if Docker crashes. In these cases, a `yb clean` fixes
things.

[Fixes ch466]