Rename biome keychain to login instead of yb
Cherry-pick of 6ec6c9dc7018205e6624e798bd4b4a1498ece8bd.

There are many tools that seem to hardcode "login" as the name of the
default keychain, and it would be better to have a default that is
compatible rather than try to be different for the sake of shaking out
code relying on implicit interfaces.

[Updates ch4145]