Port content from docs.yourbase.io
Most of the content is copied verbatim, with some light reformatting for
GitHub-Flavored Markdown. However, the List of Build Packs was
re-authored with the most up-to-date information and now includes
examples of accepted versions.

I omitted "Getting Started with YourBase" and "How to Write Build
Targets" because they are redundant with the content in README.md.
The "Remote Building" page is better served by `yb remotebuild --help`
instead of having a separate page we have to update.

[Fixes ch4338]